requirements.txt
h5py==3.6.0
matplotlib==3.5.3
numpy==1.21.6
opencv_python==4.2.0.34
pandas==1.3.5
Pillow==9.2.0
Pillow==10.3.0
scikit_learn==1.0.2
scipy==1.7.3
tensorboardX==2.0
tensorboardX==2.6.2.2
torch==1.11.0+cu113
torchvision==0.12.0+cu113
torchvision==0.6.1+cu92
tqdm==4.62.3
~atplotlib==3.1.1
~orch==1.13.1
